extend tests to save/load for attribute matrix
Add test to test_Attribute.cpp to exercise loading and saving of an
eigen matrix as an attribute.

This changed forced refactoring the test helper code that was buried in
test_MatrixRoundTrip.cpp to be moved to a header, gtest-helpers.cpp, so
attributes may be tested the same as datasets.

Modify CMakeLists.txt to work with both msvc (at least VS 2013) and
gcc (4.8)